move mapParallel tests to map tests, add sleep function for tests
